# [Issue-3367] Using serde to ser/de Aggregate function states in binary format