--- a/common/datavalues/src/types/data_df_type.rs
+++ b/common/datavalues/src/types/data_df_type.rs
@@ -15,6 +15,8 @@
 use common_arrow::arrow::types::NativeType;
 use common_io::prelude::*;
 use num::NumCast;
+use serde::de::DeserializeOwned;
+use serde::Serialize;
 
 use super::data_type::*;
 use crate::DFTryFrom;
@@ -85,6 +87,8 @@ pub trait DFPrimitiveType:
     + Default
     + BinarySer
     + BinaryDe
+    + Serialize
+    + DeserializeOwned
     + DFTryFrom<DataValue>
 {
     type LargestType: DFPrimitiveType;