--- a/common/exception/Cargo.toml
+++ b/common/exception/Cargo.toml
@@ -15,6 +15,7 @@ common-arrow = { path = "../arrow" }
 
 anyhow = "1.0.51"
 backtrace = "0.3.63"
+bincode = "1.3.3"
 octocrab = "0.15.1"
 serde = { version = "1.0.131", features = ["derive"] }
 serde_json = "1.0.72"