--- a/common/exception/src/exception.rs
+++ b/common/exception/src/exception.rs
@@ -398,6 +398,12 @@ impl From<common_arrow::arrow::error::ArrowError> for ErrorCode {
     }
 }
 
+impl From<Box<bincode::ErrorKind>> for ErrorCode {
+    fn from(error: Box<bincode::ErrorKind>) -> Self {
+        ErrorCode::from_std_error(error)
+    }
+}
+
 impl From<serde_json::Error> for ErrorCode {
     fn from(error: serde_json::Error) -> Self {
         ErrorCode::from_std_error(error)