--- a/common/functions/Cargo.toml
+++ b/common/functions/Cargo.toml
@@ -20,25 +20,26 @@ common-io = {path = "../io"}
 # Github dependencies
 
 # Crates.io dependencies
-dyn-clone = "1.0.4"
-indexmap = "1.7.0"
-lazy_static = "1.4.0"
-serde = { version = "1.0.131", features = ["derive"] }
-serde_json = "1.0.72"
+base64 = "0.13.0"
+bincode = "1.3.3"
 bytes = "1.1.0"
+crc32fast = "1.3.0"
+dyn-clone = "1.0.4"
+hex = "0.4.3"
+indexmap = "1.7.0"
+itertools = "0.10.3"
+lazy_static = "1.4.0"
+md5 = "0.7.0"
 num = "0.4.0"
 ordered-float = "2.8.0"
-crc32fast = "1.3.0"
+rand = "0.8.4"
+serde = {version = "1.0.131", features = ["derive"]}
+serde_json = "1.0.72"
 sha1 = "0.6.0"
 sha2 = "0.10.0"
-md5 = "0.7.0"
-rand = "0.8.4"
-hex = "0.4.3"
-base64 = "0.13.0"
-itertools = "0.10.3"
 
 [dev-dependencies]
 bumpalo = "3.8.0"
 common-datablocks = {path = "../datablocks"}
-pretty_assertions = "1.0.0"
 float-cmp = "0.9.0"
+pretty_assertions = "1.0.0"