--- a/common/functions/src/aggregates/aggregate_arg_min_max.rs
+++ b/common/functions/src/aggregates/aggregate_arg_min_max.rs
@@ -22,6 +22,9 @@ use common_datavalues::prelude::*;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_io::prelude::*;
+use serde::de::DeserializeOwned;
+use serde::Deserialize;
+use serde::Serialize;
 
 use super::StateAddr;
 use crate::aggregates::aggregate_function_factory::AggregateFunctionDescription;
@@ -47,11 +50,14 @@ pub trait AggregateArgMinMaxState: Send + Sync + 'static {
     fn merge_result(&mut self) -> Result<DataValue>;
 }
 
+#[derive(Serialize, Deserialize)]
 struct NumericState<T: DFPrimitiveType> {
+    #[serde(bound(deserialize = "T: DeserializeOwned"))]
     pub value: Option<T>,
     pub data: DataValue,
 }
 
+#[derive(Serialize, Deserialize)]
 struct StringState {
     pub value: Option<Vec<u8>>,
     pub data: DataValue,
@@ -148,12 +154,12 @@ where
     }
 
     fn serialize(&self, writer: &mut BytesMut) -> Result<()> {
-        self.value.serialize_to_buf(writer)?;
-        self.data.serialize_to_buf(writer)
+        let writer = BufMut::writer(writer);
+        bincode::serialize_into(writer, self)?;
+        Ok(())
     }
     fn deserialize(&mut self, reader: &mut &[u8]) -> Result<()> {
-        self.value = Option::<T>::deserialize(reader)?;
-        self.data = DataValue::deserialize(reader)?;
+        *self = bincode::deserialize_from(reader)?;
         Ok(())
     }
 
@@ -244,13 +250,17 @@ impl AggregateArgMinMaxState for StringState {
     }
 
     fn serialize(&self, writer: &mut BytesMut) -> Result<()> {
-        self.value.serialize_to_buf(writer)?;
-        self.data.serialize_to_buf(writer)
+        // self.value.serialize_to_buf(writer)?;
+        // self.data.serialize_to_buf(writer)
+        let writer = BufMut::writer(writer);
+        bincode::serialize_into(writer, self)?;
+        Ok(())
     }
 
     fn deserialize(&mut self, reader: &mut &[u8]) -> Result<()> {
-        self.value = Option::<Vec<u8>>::deserialize(reader)?;
-        self.data = DataValue::deserialize(reader)?;
+        // self.value = Option::<Vec<u8>>::deserialize(reader)?;
+        // self.data = DataValue::deserialize(reader)?;
+        *self = bincode::deserialize_from(reader)?;
         Ok(())
     }