--- a/common/functions/src/aggregates/aggregate_covariance.rs
+++ b/common/functions/src/aggregates/aggregate_covariance.rs
@@ -22,6 +22,8 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use common_io::prelude::*;
 use num::cast::AsPrimitive;
+use serde::Deserialize;
+use serde::Serialize;
 
 use super::StateAddr;
 use crate::aggregates::aggregate_function_factory::AggregateFunctionDescription;
@@ -30,6 +32,7 @@ use crate::aggregates::AggregateFunction;
 use crate::aggregates::AggregateFunctionRef;
 use crate::with_match_primitive_types;
 
+#[derive(Serialize, Deserialize)]
 pub struct AggregateCovarianceState {
     pub count: u64,
     pub co_moments: f64,
@@ -227,19 +230,14 @@ where
 
     fn serialize(&self, place: StateAddr, writer: &mut BytesMut) -> Result<()> {
         let state = place.get::<AggregateCovarianceState>();
-        state.count.serialize_to_buf(writer)?;
-        state.co_moments.serialize_to_buf(writer)?;
-        state.left_mean.serialize_to_buf(writer)?;
-        state.right_mean.serialize_to_buf(writer)?;
+        let writer = BufMut::writer(writer);
+        bincode::serialize_into(writer, state)?;
         Ok(())
     }
 
     fn deserialize(&self, place: StateAddr, reader: &mut &[u8]) -> Result<()> {
         let state = place.get::<AggregateCovarianceState>();
-        state.count = u64::deserialize(reader)?;
-        state.co_moments = f64::deserialize(reader)?;
-        state.left_mean = f64::deserialize(reader)?;
-        state.right_mean = f64::deserialize(reader)?;
+        *state = bincode::deserialize_from(reader)?;
         Ok(())
     }