--- a/common/io/src/lib.rs
+++ b/common/io/src/lib.rs
@@ -14,9 +14,7 @@
 
 pub mod prelude;
 
-mod binary_de;
 mod binary_read;
-mod binary_ser;
 mod binary_write;
 mod buf_read;
 mod marshal;