I hereby agree to the terms of the CLA available at: https://databend.rs/policies/cla/

## Summary

1. Using serde to ser/de Aggregate function states in binary format.
2. Clean up non used BinarySer/BinaryDe code.

## Changelog

- Improvement
- Performance Improvement

## Related Issues

Fixes #3367

## Test Plan

Unit Tests

Stateless Tests