@BohuTANG 2021-12-14
Benchmark test in my local PC(Without performance degradation :+1: ):

This patch:
```
databend :) SELECT sum(number) FROM numbers_mt(100000000000);

┌────────sum(number)─┐
│ 932355974711512064 │
└────────────────────┘

1 rows in set. Elapsed: 4.830 sec. Processed 100.00 billion rows, 800.00 GB (20.70 billion rows/s., 165.62 GB/s.)

databend :) SELECT min(number) FROM numbers_mt(100000000000);

┌─min(number)─┐
│           0 │
└─────────────┘

1 rows in set. Elapsed: 9.606 sec. Processed 100.00 billion rows, 800.00 GB (10.41 billion rows/s., 83.28 GB/s.)

databend :) SELECT max(number), sum(number) FROM numbers_mt(1000000000) GROUP BY number % 3, number % 4, number % 5 LIMIT 10;

┌─max(number)─┬──────sum(number)─┐
│   999999992 │ 8333333700000004 │
│   999999957 │ 8333333116666662 │
│   999999983 │ 8333333550000001 │
│   999999986 │ 8333333600000002 │
│   999999981 │ 8333333516666667 │
│   999999993 │ 8333333716666671 │
│   999999949 │ 8333332983333334 │
│   999999958 │ 8333333133333328 │
│   999999945 │ 8333332916666670 │
│   999999973 │ 8333333383333331 │
└─────────────┴──────────────────┘

10 rows in set. Elapsed: 1.654 sec. Processed 1.00 billion rows, 8.00 GB (604.69 million rows/s., 4.84 GB/s.)

```


main branch:
```
databend :) SELECT sum(number) FROM numbers_mt(100000000000);


┌────────sum(number)─┐
│ 932355974711512064 │
└────────────────────┘

1 rows in set. Elapsed: 5.126 sec. Processed 100.00 billion rows, 800.00 GB (19.51 billion rows/s., 156.08 GB/s.)

databend :) SELECT min(number) FROM numbers_mt(100000000000);


┌─min(number)─┐
│           0 │
└─────────────┘

1 rows in set. Elapsed: 9.509 sec. 

databend :) SELECT max(number), sum(number) FROM numbers_mt(1000000000) GROUP BY number % 3, number % 4, number % 5 LIMIT 10;


┌─max(number)─┬──────sum(number)─┐
│   999999940 │ 8333332833333340 │
│   999999965 │ 8333333249999995 │
│   999999973 │ 8333333383333331 │
│   999999981 │ 8333333516666667 │
│   999999997 │ 8333333783333339 │
│   999999955 │ 8333333083333330 │
│   999999960 │ 8333333166666660 │
│   999999967 │ 8333333283333329 │
│   999999977 │ 8333333449999999 │
│   999999948 │ 8333332966666668 │
└─────────────┴──────────────────┘

10 rows in set. Elapsed: 1.656 sec. Processed 1.00 billion rows, 8.00 GB (603.97 million rows/s., 4.83 GB/s.)
```