@sundy-li 2021-12-14
Can we optimize the length encoding to variant encoding?

 https://github.com/bincode-org/bincode/blob/trunk/src/enc/mod.rs#L68-L70