@ygf11 2021-12-14
> Can we optimize the length encoding to variant encoding?

Yes, it supports this option. Let me do it.

https://github.com/bincode-org/bincode/blob/c44b5e364e7084cdbabf9f94b63a3c7f32b8fb68/src/config/mod.rs#L156-L159