@sundy-li 2021-12-14
Yes, but I don't want it to be a global setting.

Could it be possible to just write the `offsets`/ `length`   of (vectors,  byte arrays ) in variant encoding?
For other integers, we just use native encoding.

This is because we always know that the vector state has small size ranges, so it benefits a lot to use variant encoding. But for other integers, the range may be random, it's better to use native encoding.