@ygf11 2021-12-14
> Yes, but I don't want it to be a global setting.

We can serialize every field in state one by one, like:
```
struct State {
   v1: Vec<u8>,
   v2: u8,
   v3: u16   
}

fn serialize(&self, writer: &mut ByteMut) -> Result<()> {
   let writer: std::io::Writer = BufMut::writer(writer);

   bincode_serialize_variant_encoding(writer, self.v1);
   bincode_serialize_fixed_encoding(writer, self.v2);
   bincode_serialize_fixed_encoding(writer, self.v3);
   Ok(())
}
```


the problem is that `serialize_into` method need the ownership of the `writer`, so we can not write one by one here.