@ygf11 2021-12-14
I also find the `2.0.0-alpha.1` of `bincode` crate support passing `&mut [u8]` as writer, but it is still in developing.