@sundy-li 2021-12-14
How about using a forked `bincode` crate ?