@ygf11 2021-12-14
> How about using a forked bincode crate ?

Current `bincode` is enough, I misunderstand the meaning of api :sweat: